requirements.txt
numpy
scipy
cached_property
lacecore[obj] @ git+https://github.com/lace/lacecore.git@reindex-vertices
click>=7.0,<8.0
tqdm
polliwog==1.0.0b7
pyyaml>=5.1
